# docs/openapi-modules/server-api.openapi.yaml
openapi: 3.0.3
info:
  title: PcManager — Server.Api
  description: |
    Модуль **Server**: REST API процесса Server.Api.

    Импорт в Apidog: «Импорт» → OpenAPI → этот файл → сохранить как модуль **Server**
    (или объединить с другими модулями проекта). Base URL по умолчанию `http://127.0.0.1:8080`.

    Ответы CommandExecutionResponseDto: поля `type` и `status` — числа (enum).
  version: 1.0.0

servers:
  - url: "{baseUrl}"
    description: Server.Api
    variables:
      baseUrl:
        default: http://127.0.0.1:8080

tags:
  - name: Auth
    description: JWT
  - name: Commands
    description: Команды
  - name: Notifications
    description: Оповещения
  - name: System
    description: Система

paths:
  /api/auth/token:
    post:
      tags: [Auth]
      summary: Получить JWT
      operationId: serverAuthIssueToken
      parameters:
        - name: role
          in: query
          required: false
          schema:
            type: string
            default: Operator
            enum: [Operator, Admin]
      responses:
        "200":
          description: Строка JWT
          content:
            application/json:
              schema:
                type: string

  /api/commands/execute:
    post:
      tags: [Commands]
      summary: Выполнить команду
      operationId: serverCommandsExecute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteCommandRequestDto"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandExecutionResponseDto"
        "401":
          description: Unauthorized

  /api/commands/{id}:
    get:
      tags: [Commands]
      summary: Статус команды
      operationId: serverCommandsGetStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandExecutionResponseDto"
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /api/notifications/inbound:
    post:
      tags: [Notifications]
      summary: Входящее оповещение (webhook)
      description: Заголовок X-Notification-Key = Notifications:InboundApiKey. Без JWT.
      operationId: serverNotificationsInbound
      security:
        - notificationKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboundNotificationRequestDto"
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
        "503":
          description: InboundApiKey не задан на сервере

  /api/notifications/recent:
    get:
      tags: [Notifications]
      summary: Оповещения после sinceUtc
      operationId: serverNotificationsRecent
      security:
        - bearerAuth: []
      parameters:
        - name: sinceUtc
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationBroadcastDto"
        "401":
          description: Unauthorized

  /api/system/info:
    get:
      tags: [System]
      summary: Информация о системе
      operationId: serverSystemInfo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfoDto"
        "401":
          description: Unauthorized

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    notificationKey:
      type: apiKey
      in: header
      name: X-Notification-Key

  schemas:
    ExecuteCommandRequestDto:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [LockWorkstation, GetUptime]
        payload:
          type: string
          nullable: true

    CommandExecutionResponseDto:
      type: object
      properties:
        commandId:
          type: string
          format: uuid
        type:
          type: integer
          enum: [1, 2]
        status:
          type: integer
          enum: [1, 2, 3, 4]
        result:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    InboundNotificationRequestDto:
      type: object
      required: [title]
      properties:
        title:
          type: string
        message:
          type: string
          nullable: true

    NotificationBroadcastDto:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
          nullable: true
        receivedAtUtc:
          type: string
          format: date-time

    SystemInfoDto:
      type: object
      properties:
        machineName:
          type: string
        osDescription:
          type: string
        framework:
          type: string
        utcNow:
          type: string
          format: date-time